+++ .idea/modules/scala-uml.iml
@@ -33,7 +33,17 @@
     <orderEntry type="library" name="sbt: com.lihaoyi:sourcecode_2.13:0.2.1:jar" level="project" />
     <orderEntry type="library" name="sbt: com.thesamet.scalapb:lenses_2.13:0.10.9:jar" level="project" />
     <orderEntry type="library" name="sbt: com.thesamet.scalapb:scalapb-runtime_2.13:0.10.9:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-actor-typed_2.13:2.6.18:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-actor_2.13:2.6.18:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-http-core_2.13:10.2.7:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-http-spray-json_2.13:10.2.7:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-http_2.13:10.2.7:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-parsing_2.13:10.2.7:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-protobuf-v3_2.13:2.6.18:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-slf4j_2.13:2.6.18:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe.akka:akka-stream_2.13:2.6.18:jar" level="project" />
     <orderEntry type="library" name="sbt: com.typesafe:config:1.4.0:jar" level="project" />
+    <orderEntry type="library" name="sbt: com.typesafe:ssl-config-core_2.13:0.4.2:jar" level="project" />
     <orderEntry type="library" name="sbt: io.chrisdavenport:unique_2.13:2.0.0:jar" level="project" />
     <orderEntry type="library" name="sbt: io.chrisdavenport:vault_2.13:2.0.0:jar" level="project" />
     <orderEntry type="library" name="sbt: io.circe:circe-core_2.13:0.13.0:jar" level="project" />
@@ -41,6 +51,7 @@
     <orderEntry type="library" name="sbt: io.circe:circe-jawn_2.13:0.13.0:jar" level="project" />
     <orderEntry type="library" name="sbt: io.circe:circe-literal_2.13:0.13.0:jar" level="project" />
     <orderEntry type="library" name="sbt: io.circe:circe-numbers_2.13:0.13.0:jar" level="project" />
+    <orderEntry type="library" name="sbt: io.spray:spray-json_2.13:1.3.6:jar" level="project" />
     <orderEntry type="library" name="sbt: net.java.dev.jna:jna:5.3.1:jar" level="project" />
     <orderEntry type="library" name="sbt: net.sourceforge.plantuml:plantuml:8059:jar" level="project" />
     <orderEntry type="library" name="sbt: org.bitbucket.inkytonik.dsinfo:dsinfo_2.13:0.4.0:jar" level="project" />
@@ -54,7 +65,10 @@
     <orderEntry type="library" name="sbt: org.http4s:parboiled_2.13:2.0.1:jar" level="project" />
     <orderEntry type="library" name="sbt: org.jline:jline:3.15.0:jar" level="project" />
     <orderEntry type="library" name="sbt: org.log4s:log4s_2.13:1.8.2:jar" level="project" />
+    <orderEntry type="library" name="sbt: org.reactivestreams:reactive-streams:1.0.3:jar" level="project" />
     <orderEntry type="library" name="sbt: org.scala-lang.modules:scala-collection-compat_2.13:2.1.6:jar" level="project" />
+    <orderEntry type="library" name="sbt: org.scala-lang.modules:scala-java8-compat_2.13:1.0.0:jar" level="project" />
+    <orderEntry type="library" name="sbt: org.scala-lang.modules:scala-parser-combinators_2.13:1.1.2:jar" level="project" />
     <orderEntry type="library" scope="TEST" name="sbt: org.scala-lang.modules:scala-xml_2.13:1.2.0:jar" level="project" />
     <orderEntry type="library" name="sbt: org.scala-lang:scala-compiler:2.13.3:jar" level="project" />
     <orderEntry type="library" name="sbt: org.scala-lang:scala-library:2.13.3:jar" level="project" />
@@ -84,7 +98,7 @@
     <orderEntry type="library" scope="TEST" name="sbt: org.scalatest:scalatest_2.13:3.2.2:jar" level="project" />
     <orderEntry type="library" name="sbt: org.scodec:scodec-bits_2.13:1.1.18:jar" level="project" />
     <orderEntry type="library" name="sbt: org.scoverage:scalac-scoverage-runtime_2.13:1.4.1:jar" level="project" />
-    <orderEntry type="library" name="sbt: org.slf4j:slf4j-api:1.7.25:jar" level="project" />
+    <orderEntry type="library" name="sbt: org.slf4j:slf4j-api:1.7.32:jar" level="project" />
     <orderEntry type="library" name="sbt: org.typelevel:cats-core_2.13:2.2.0:jar" level="project" />
     <orderEntry type="library" name="sbt: org.typelevel:cats-effect_2.13:2.2.0:jar" level="project" />
     <orderEntry type="library" name="sbt: org.typelevel:cats-kernel_2.13:2.2.0:jar" level="project" />
