+++ .idea/workspace.xml
@@ -5,11 +5,14 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="b894318d-1f1b-4c84-a1aa-11290f7e44e1" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/src/test/scala/collector/TypeCollectorSpec.scala" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/scala/app/ci/umlProcessing.scala" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/scala/rest/ASTCreator.scala" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/scala/rest/ASTRoutes.scala" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/modules/scala-uml-build.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/modules/scala-uml-build.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/modules/scala-uml.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/modules/scala-uml.iml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/scala/scalameta/types/TargetTypeCollector.scala" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/scala/scalameta/types/TargetTypeCollector.scala" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/scala/scalameta/util/context/LocalContext.scala" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/scala/scalameta/util/context/LocalContext.scala" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/scala/uml/ast.scala" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/scala/uml/ast.scala" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/build.sbt" beforeDir="false" afterPath="$PROJECT_DIR$/build.sbt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/scala/app/ci/Main.scala" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/scala/app/ci/Main.scala" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -107,7 +110,7 @@
       <recent name="assets" />
     </key>
   </component>
-  <component name="RunManager" selected="ScalaTest.TypeCollectorSpec.Template Type with depth n should be transformed correctly into a RefTemplate">
+  <component name="RunManager" selected="ScalaTest.TypeCollectorSpec">
     <configuration default="true" type="ArquillianJUnit" factoryName="" nameIsGenerated="true">
       <option name="arquillianRunConfiguration">
         <value>
@@ -119,16 +122,16 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="Main (1)" type="Application" factoryName="Application" temporary="true">
-      <option name="MAIN_CLASS_NAME" value="app.ci.Main" />
+    <configuration name="TypeCollectorSpec" type="ScalaTestRunConfiguration" factoryName="ScalaTest" temporary="true" nameIsGenerated="true" show_console_on_std_err="false" show_console_on_std_out="false">
       <module name="scala-uml" />
-      <option name="PROGRAM_PARAMETERS" value="-f $PROJECT_DIR$/src/main/scala/app/ci -o $PROJECT_DIR$/../../ -t" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="app.ci.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
+      <option name="allowRunningInParallel" value="false" />
+      <option name="projectPathOnTarget" />
+      <option name="selectedOptions">
+        <list />
+      </option>
+      <option name="testKind" value="Class" />
+      <option name="testClassPath" value="collector.TypeCollectorSpec" />
+      <option name="workingDirectory" value="C:\Users\dbs476\Documents\Repos\scala-uml" />
       <method v="2">
         <option name="Make" enabled="true" />
       </method>
@@ -194,19 +197,19 @@
       </method>
     </configuration>
     <list>
-      <item itemvalue="Application.Main (1)" />
       <item itemvalue="ScalaTest.TypeCollectorSpec.A Type of the form Type.Select should be transformed into a RefPathQualifier with correct order" />
       <item itemvalue="ScalaTest.TypeCollectorSpec.Single Template type should be transformed properly" />
       <item itemvalue="ScalaTest.TypeCollectorSpec.Map template type should be transformed into a RefTemplate with two type parameters" />
       <item itemvalue="ScalaTest.TypeCollectorSpec.Template Type with depth n should be transformed correctly into a RefTemplate" />
+      <item itemvalue="ScalaTest.TypeCollectorSpec" />
     </list>
     <recent_temporary>
       <list>
+        <item itemvalue="ScalaTest.TypeCollectorSpec" />
         <item itemvalue="ScalaTest.TypeCollectorSpec.Template Type with depth n should be transformed correctly into a RefTemplate" />
         <item itemvalue="ScalaTest.TypeCollectorSpec.Map template type should be transformed into a RefTemplate with two type parameters" />
         <item itemvalue="ScalaTest.TypeCollectorSpec.Single Template type should be transformed properly" />
         <item itemvalue="ScalaTest.TypeCollectorSpec.A Type of the form Type.Select should be transformed into a RefPathQualifier with correct order" />
-        <item itemvalue="Application.Main (1)" />
       </list>
     </recent_temporary>
   </component>
@@ -349,6 +352,8 @@
       <workItem from="1639757664011" duration="8501000" />
       <workItem from="1640123053151" duration="2558000" />
       <workItem from="1640729090548" duration="16216000" />
+      <workItem from="1642457411087" duration="2030000" />
+      <workItem from="1642688600625" duration="5506000" />
     </task>
     <servers />
   </component>
