+++ src/main/scala/rest/ASTRoutes.scala
@@ -1,0 +1,5 @@
+package rest
+
+class ASTRoutes() {
+
+}
